Move pull test part into pull
Towards #290
